hipaa_compliance_tool/requirements.txt
# Core dependencies
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=0.19.0

# Security and encryption
cryptography>=3.4.8
bcrypt>=3.2.0
PyJWT>=2.4.0

# Email
secure-smtplib>=0.1.1

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.8.0

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional monitoring (uncomment if using)
# sentry-sdk>=1.9.0

# Database (if using SQLite/PostgreSQL)
# sqlalchemy>=1.4.0
# psycopg2-binary>=2.9.0  # for PostgreSQL

# Web framework (if adding web interface)
# flask>=2.2.0
# flask-login>=0.6.0
# flask-wtf>=1.0.0
